Split off Angle0to2pi into its own header file
