+++ DataFormats/Math/interface/deltaPhi.h
@@ -99,14 +99,15 @@ namespace angle_units {
   }
 }
 
+
 namespace angle0to2pi {
 
   using namespace angle_units::operators;
 
   // make0to2pi constrains an angle to be >= 0 and < 2pi.
-	// This function is a faster version of reco::reduceRange.
-	// In timing tests, it takes about half the time of reco::reduceRange.
-	// It also protects against floating-point value drift over repeated calculations.
+  // This function is a faster version of reco::reduceRange.
+  // In some timing tests, it can take about half the time of reco::reduceRange.
+  // It also protects against floating-point value drift over repeated calculations.
 
   template <class valType>
   inline constexpr valType make0to2pi(valType angle) {
@@ -127,64 +128,6 @@ namespace angle0to2pi {
     }
     return (angle);
   }
-
-
-  // Angle0to2pi class keep its  value as >= 0 and < 2pi.
-  //
-  // nearZero() tells whether the angle is close enough to 0 to be considered 0.
-  // The default tolerance is 1 degree.
-  //
-  // degrees() returns the angle in degrees.
-
-  template <class valType>
-  class Angle0to2pi {
-  public:
-
-    Angle0to2pi(valType ang) :
-    ang0to2pi (make0to2pi(ang))
-    {
-    }
-
-    inline operator valType() const {
-      return (ang0to2pi);
-    }
-
-    inline bool nearZero(float tolerance = 1.0_deg) const {
-      return (std::abs(ang0to2pi) - tolerance <= 0.0);
-    }
-
-    inline bool nearEqual(const Angle0to2pi<valType> &angle, float tolerance = 0.001) const {
-      return (std::abs(ang0to2pi - angle) - tolerance <= 0.0);
-    }
-
-    inline valType degrees() const {
-      return (convertRadToDeg(ang0to2pi));
-    }
-
-    inline Angle0to2pi & operator +=(valType ang) {
-      ang0to2pi = make0to2pi(ang0to2pi + ang);
-      return (*this);
-    }
-
-    inline Angle0to2pi & operator -=(valType ang) {
-      ang0to2pi = make0to2pi(ang0to2pi - ang );
-      return (*this);
-    }
-
-    inline Angle0to2pi & operator *=(valType val) {
-      ang0to2pi = make0to2pi(ang0to2pi * val);
-      return (*this);
-    }
-
-    inline Angle0to2pi & operator /=(valType val) {
-      ang0to2pi = make0to2pi(ang0to2pi / val);
-      return (*this);
-    }
-
-  private:
-    valType ang0to2pi;
-  };
-
 }
 
 #endif
